# openapi/ku-seek-api.yml
openapi: 3.0.0
info:
  title: KU SEEK
  description: CRUD operations for KU SEEK
  version: 2.0.0
servers:
  - url: /api/v1
paths:
...
  /refresh:
    get:
      tags:
        - Authentication
      summary: Get a new access token from a refesh token
      operationId: controllers.auth_controller.get_new_access_token
      parameters:
        - in: cookie
          name: refresh_token
          schema:
            type: string
          required: true
      responses:
        200:
          description: Return a new access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
        400:
          description: Invalid refresh token provided in header
...
  /auth/credentials:
    post:
      tags:
          - Authentication
      summary: Login a user using credentials and return user credentials.
      operationId: controllers.auth_controller.handle_credential_login
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: The user's email
                password:
                  type: string
                  description: The user's password
      responses: 
        200:
          description: return client_id and access/refresh tokens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserCredentials'
        400:
          description: Invalid authorization code
  /auth/oauth:
    post:
      tags:
        - Authentication
      summary: Login or Register a user and return user credentials.
      operationId: controllers.auth_controller.handle_authentication
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                code:
                  type: string
                  description: The grant code from the oauth2 flow
      responses:
        200:
          description: return client_id and access/refresh tokens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserCredentials"
        400:
          description: Invalid authorization code
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserCredentials:
      type: object
      required:
        - access_token
      properties:
        access_token:
          type: string
          description: The access token
        email:
          description: The account email
        type:
          description: The account type
        user_id:
          description: The user id
...